+++ lib/ui/groups/GroupCreatePanel.dart
@@ -20,6 +20,7 @@ import 'dart:io';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/ext/Group.dart';
+import 'package:illinois/ext/ImagesResult.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/ui/groups/GroupAdvancedSettingsPanel.dart';
 import 'package:illinois/ui/attributes/ContentAttributesPanel.dart';
@@ -304,15 +305,12 @@ class _GroupCreatePanelState extends State<GroupCreatePanel> {
 
   void _onTapAddImage() async {
     Analytics().logSelect(target: "Add Image");
-    String? _imageUrl = await GroupAddImageWidget.show(context: context, updateUrl: _group!.imageURL);
-    if (_imageUrl != null) {
-      if (mounted) {
-        setState(() {
-          _group!.imageURL = _imageUrl;
-        });
-      }
-    }
-    Log.d("Image Url: $_imageUrl");
+    ImagesResult? result = await GroupAddImageWidget.show(context: context, url: _group!.imageURL).then((result) => result);
+    if(result?.succeeded == true)
+    setStateIfMounted(() {
+      _group!.imageURL = result?.stringData;
+    });
+    Log.d("Image Url: ${result?.stringData}");
   }
 
   //Name