+++ lib/ui/settings/SettingsHomeContentPanel.dart
@@ -160,7 +160,7 @@ class _SettingsHomeContentPanelState extends State<SettingsHomeContentPanel> imp
             Semantics( label: Localization().getStringEx('dialog.close.title', 'Close'), hint: Localization().getStringEx('dialog.close.hint', ''), inMutuallyExclusiveGroup: true, button: true, child:
               InkWell(onTap : _onTapClose, child:
                 Container(padding: EdgeInsets.only(left: 8, right: 16, top: 16, bottom: 16), child:
-                  Styles().images.getImage('close', excludeFromSemantics: true),
+                  Styles().images.getImage('close-circle', excludeFromSemantics: true),
                 ),
               ),
             ),